chore: rename modules (#89)
