+++ CHANGELOG.md
@@ -1,3 +1,15 @@
 # @biconomy/sdk
 
+## 0.0.4
+
+### Patch Changes
+
+- renamed validator modules
+
+## 0.0.3
+
+### Patch Changes
+
+- modules dx improvements
+
 ## 0.0.0
