+++ README.md
@@ -41,22 +41,17 @@ bun i
 
 ```typescript
 import { createNexusClient } from "@biconomy/sdk";
-import { http, publicClient } from "viem";
-
-const publicClient = createPublicClient({
-  chain: mainnet,
-  transport: http(),
-});
+import { http } from "viem";
 
 const nexusClient = await createNexusClient({
-  holder: account,
+  signer: account,
   chain,
   transport: http(),
   bundlerTransport: http(bundlerUrl),
 });
 
 const hash = await nexusClient.sendTransaction({ calls: [to: "0x...", value: 1] });
-const { status, transactionHash } = await publicClient.waitForTransactionReceipt({ hash });
+const { status, transactionHash } = await nexusClient.waitForTransactionReceipt({ hash });
 
 ```
