+++ package.json
@@ -8,7 +8,7 @@
   "sideEffects": false,
   "name": "@biconomy/sdk",
   "author": "Biconomy",
-  "version": "0.0.0",
+  "version": "0.0.4",
   "description": "SDK for Biconomy integration with support for account abstraction, smart accounts, ERC-4337.",
   "keywords": [
     "erc-7579",
