+++ src/sdk/clients/createNexusSessionClient.test.ts
@@ -21,13 +21,13 @@ import {
 } from "../../test/testUtils"
 import type { MasterClient, NetworkConfig } from "../../test/testUtils"
 import addresses from "../__contracts/addresses"
-import { isSessionEnabled } from "../modules/smartSessions/Helpers"
-import type { CreateSessionDataParams } from "../modules/smartSessions/Types"
+import { isSessionEnabled } from "../modules/smartSessionsValidator/Helpers"
+import type { CreateSessionDataParams } from "../modules/smartSessionsValidator/Types"
 import {
   smartSessionCreateActions,
   smartSessionUseActions
-} from "../modules/smartSessions/decorators"
-import { toSmartSessions } from "../modules/smartSessions/toSmartSessions"
+} from "../modules/smartSessionsValidator/decorators"
+import { toSmartSessionsValidator } from "../modules/smartSessionsValidator/toSmartSessionsValidator"
 import type { Module } from "../modules/utils/Types"
 import { type NexusClient, createNexusClient } from "./createNexusClient"
 import { createNexusSessionClient } from "./createNexusSessionClient"
@@ -67,7 +67,7 @@ describe("nexus.session.client", async () => {
     })
     nexusAccountAddress = await nexusClient.account.getCounterFactualAddress()
 
-    sessionsModule = toSmartSessions({
+    sessionsModule = toSmartSessionsValidator({
       account: nexusClient.account,
       signer: eoaAccount
     })
@@ -171,7 +171,7 @@ describe("nexus.session.client", async () => {
       bundlerTransport: http(bundlerUrl)
     })
 
-    const useSessionsModule = toSmartSessions({
+    const useSessionsModule = toSmartSessionsValidator({
       account: smartSessionNexusClient.account,
       signer: sessionKeyAccount,
       moduleData: {
@@ -215,7 +215,7 @@ describe("nexus.session.client", async () => {
   }, 60000)
 
   test("session signer is not allowed to send unauthorised action", async () => {
-    const useSessionsModule = toSmartSessions({
+    const useSessionsModule = toSmartSessionsValidator({
       account: nexusClient.account,
       signer: sessionKeyAccount,
       moduleData: {
