+++ src/sdk/modules/index.ts
@@ -2,6 +2,6 @@ export * from "./utils/Types"
 export * from "./utils/Constants"
 export * from "./utils/Helpers"
 
-export * from "./k1/toK1"
-export * from "./ownables/toOwnables"
-export * from "./smartSessions/toSmartSessions"
+export * from "./k1Validator/toK1Validator"
+export * from "./ownableValidator/toOwnableValidator"
+export * from "./smartSessionsValidator/toSmartSessionsValidator"
