+++ src/sdk/modules/ownableValidator/decorators/getRemoveOwnerTx.ts
