+++ src/sdk/modules/smartSessionsValidator/Helpers.ts
