+++ src/sdk/modules/smartSessionsValidator/decorators/smartSessions.decorators.test.ts
@@ -23,7 +23,7 @@ import {
 } from "../../../clients/createNexusClient"
 import { createNexusSessionClient } from "../../../clients/createNexusSessionClient"
 import type { Module } from "../../utils/Types"
-import { toSmartSessions } from "../toSmartSessions"
+import { toSmartSessionsValidator } from "../toSmartSessionsValidator"
 import { smartSessionCreateActions, smartSessionUseActions } from "./"
 
 describe("modules.smartSessions.decorators", async () => {
@@ -67,7 +67,7 @@ describe("modules.smartSessions.decorators", async () => {
   })
 
   test("should test create smart session decorators", async () => {
-    const sessionsModule = toSmartSessions({
+    const sessionsModule = toSmartSessionsValidator({
       account: nexusClient.account,
       signer: eoaAccount
     })
@@ -89,7 +89,7 @@ describe("modules.smartSessions.decorators", async () => {
   })
 
   test("should test use smart session decorators", async () => {
-    const useSessionsModule = toSmartSessions({
+    const useSessionsModule = toSmartSessionsValidator({
       account: nexusClient.account,
       signer: sessionKeyAccount,
       moduleData: {
